Rename PR2-related collision test files (#1856)
