+++ moveit_core/collision_detection_fcl/include/moveit/collision_detection_fcl/collision_common.h
@@ -32,7 +32,7 @@
  *  POSSIBILITY OF SUCH DAMAGE.
  *********************************************************************/
 
-/* Author: Ioan Sucan, Jens Petit */
+/* Author: Ioan Sucan */
 
 #pragma once
